Dynamic Selected
Updated the genre so that when an option is selected the subgenre
options are populated. That is used to come up with movies later on.

There are still some possible bugs. Some of the subgenres weren’t put
under a genre category. The SQL command that populates the list of
results should be modified so that all of the subgenre types can be
searched for.