Search Functionality
Made it so that you can search for movie’s by title. I have not changed
all of the other pages, so if we like this, I’d ask that someone else
just put this code in the other pages. It might actually make more
sense for this to link to a search page, which someone could also do
pretty easily.